pyproject.toml
[project]
name = "stir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.6",
    "pydantic>=2.10.4",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "selenium>=4.27.1",
    "uvicorn>=0.34.0",
]
